Create a hierarchy

This functions allows to generate a hierarchical data structure that can be used in other packages such as cellKey or sdcTable.

hier_create(root = "Total", nodes = NULL)

Arguments

root
(character) name of the overall total

nodes
(character) name of leaves (nodes) in the hierarchy

Value

a (nested) sdc hierarchy tree

See also

Examples

# without nodes
h <- hier_create(root = "tot")
hier_display(h)

#> tot

# with nodes
h <- hier_create(root = "tot", nodes = LETTERS[1:5])
hier_display(h)

#> tot
#> ├─A
#> ├─B
#> ├─C
#> ├─D
#> └─E